Use profile in tests
