+++ cmd/profiles.go
@@ -8,6 +8,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
+	"strings"
 	"time"
 
 	"github.com/olekukonko/tablewriter"
@@ -18,6 +19,7 @@ import (
 	"github.com/elastic/elastic-package/internal/configuration/locations"
 	"github.com/elastic/elastic-package/internal/install"
 	"github.com/elastic/elastic-package/internal/profile"
+	"github.com/elastic/elastic-package/internal/stack"
 )
 
 // jsonFormat is the format for JSON output
@@ -28,8 +30,8 @@ const tableFormat = "table"
 
 func setupProfilesCommand() *cobraext.Command {
 	profilesLongDescription := `Use this command to add, remove, and manage multiple config profiles.
-	
-Individual user profiles appear in ~/.elastic-package/stack, and contain all the config files needed by the "stack" subcommand. 
+
+Individual user profiles appear in ~/.elastic-package/stack, and contain all the config files needed by the "stack" subcommand.
 Once a new profile is created, it can be specified with the -p flag, or the ELASTIC_PACKAGE_PROFILE environment variable.
 User profiles can be configured with a "config.yml" file in the profile directory.`
 
@@ -234,4 +236,58 @@ func availableProfilesAsAList() ([]string, error) {
 	}
 
 	return profileNames, nil
+}
+
+func getProfileFlag(cmd *cobra.Command) (*profile.Profile, error) {
+	profileName, err := cmd.Flags().GetString(cobraext.ProfileFlagName)
+	if err != nil {
+		return nil, cobraext.FlagParsingError(err, cobraext.ProfileFlagName)
+	}
+	if profileName == "" {
+		config, err := install.Configuration()
+		if err != nil {
+			return nil, fmt.Errorf("cannot read configuration: %w", err)
+		}
+		profileName = config.CurrentProfile()
+	}
+
+	p, err := profile.LoadProfile(profileName)
+	if errors.Is(err, profile.ErrNotAProfile) {
+		list, err := availableProfilesAsAList()
+		if err != nil {
+			return nil, errors.Wrap(err, "error listing known profiles")
+		}
+		if len(list) == 0 {
+			return nil, fmt.Errorf("%s is not a valid profile", profileName)
+		}
+		return nil, fmt.Errorf("%s is not a valid profile, known profiles are: %s", profileName, strings.Join(list, ", "))
+	}
+	if err != nil {
+		return nil, errors.Wrap(err, "error loading profile")
+	}
+
+	return p, nil
+}
+
+func getProviderFromProfile(cmd *cobra.Command, profile *profile.Profile, checkFlag bool) (stack.Provider, error) {
+	var providerName = stack.DefaultProvider
+	stackConfig, err := stack.LoadConfig(profile)
+	if err != nil {
+		return nil, err
+	}
+	if stackConfig.Provider != "" {
+		providerName = stackConfig.Provider
+	}
+
+	if checkFlag {
+		providerFlag, err := cmd.Flags().GetString(cobraext.StackProviderFlagName)
+		if err != nil {
+			return nil, cobraext.FlagParsingError(err, cobraext.StackProviderFlagName)
+		}
+		if providerFlag != "" {
+			providerName = providerFlag
+		}
+	}
+
+	return stack.BuildProvider(providerName, profile)
 }
