+++ internal/testrunner/runners/system/runner.go
@@ -7,6 +7,8 @@ package system
 import (
 	"encoding/json"
 	"fmt"
+	"io/fs"
+	"log"
 	"math/rand"
 	"os"
 	"path/filepath"
@@ -24,6 +26,7 @@ import (
 	"github.com/elastic/elastic-package/internal/multierror"
 	"github.com/elastic/elastic-package/internal/packages"
 	"github.com/elastic/elastic-package/internal/signal"
+	"github.com/elastic/elastic-package/internal/stack"
 	"github.com/elastic/elastic-package/internal/testrunner"
 	"github.com/elastic/elastic-package/internal/testrunner/runners/system/servicedeployer"
 )
@@ -182,6 +185,46 @@ func (r *runner) run() (results []testrunner.TestResult, err error) {
 			}
 		}
 	}
+
+	// check agent logs
+	tempDir, err := os.MkdirTemp("", "test-system-")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer os.RemoveAll(tempDir)
+
+	dumpFolder, err := stack.Dump(stack.DumpOptions{
+		Output:  tempDir,
+		Profile: r.options.Profile,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// logs from elastic-agent >8.6.0
+	// stack.ParseErrorLogs()
+	containers := []string{"elastic_agent"}
+	for _, c := range containers {
+		elasticAgentLogsFolder := filepath.Join(dumpFolder, "logs", "elastic-agent-internal")
+
+		err = filepath.WalkDir(elasticAgentLogsFolder, func(path string, d fs.DirEntry, err error) error {
+			if err != nil {
+				return err
+			}
+			if d.IsDir() {
+				return nil
+			}
+
+			logger.Debugf("Checking file as elastic-agent log: %s", path)
+
+			// read file and look for errors
+			return nil
+		})
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	return results, nil
 }
